pyproject.toml
[tool.poetry]
name = "repl-nix-nighttradervanillajs"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
psycopg2-binary = "^2.9.9"
flask-sqlalchemy = "^3.1.1"
pyjwt = "^2.9.0"
flask-socketio = "^5.3.7"
pyotp = "^2.9.0"
qrcode = "^7.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
